Make small functions `static inline` and move them to header files
